Refactor Livebook.Hubs API
