--- lib/livebook/hub_provider.ex
@@ -1,9 +1,0 @@
-defprotocol Livebook.HubProvider do
-  @moduledoc false
-
-  @doc """
-  Gets a list of hubs.
-  """
-  @spec fetch_hubs(any()) :: {:ok, list(Livebook.HubProvider.Hub.t())} | {:error, any()}
-  def fetch_hubs(hub)
-end
